feat: Add Configuration to readme.md (#361)
This may be obvious to seasoned HA users, but it took me a long time to
work out that after installing the Solarman integration I still needed a
separate step to add it under Integrations (especially since
autodiscovery is mentioned in various places). I propose adding the
instructions to `readme.md`, but they would also work fine in the wiki.