Update max range of "Zero Export power" to allow negative values
By using a SUN-12K-SG04LP3-EU inverter with firmware 1135 it's possible to use negative zero export power. For example to set the zero export power to -20W we can set it to 65516. So for using negative 'Zero Export power' we need to have a higher limit for that setting.